third_party/WebKit/LayoutTests/inspector/animation/animation-timeline-fill.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>
var animation;

function startAnimationWithFillBackwards()
{
    if (animation)
        animation.cancel();
    animation = node.animate([{ width: "100px" }, { width: "200px" }], { duration: 200, delay: 100, endDelay: 100, fill: "backwards" });
}

function startAnimationWithFillForwards()
{
    if (animation)
       animation.cancel();
    animation = node.animate([{ width: "100px" }, { width: "200px" }], { duration: 200, delay: 100, endDelay: 100, fill: "forwards" });
}

function startAnimationWithFillBoth()
{
    if (animation)
        animation.cancel();
    animation = node.animate([{ width: "100px" }, { width: "200px" }], { duration: 200, delay: 100, endDelay: 100, fill: "both" });
}

var initialize_Animations = function() {

    InspectorTest.preloadModule("animation");
}

function test()
{
    // Override timeline width for testing
    WebInspector.AnimationTimeline.prototype.width = function() { return 1000; }
    // Override animation color for testing
    WebInspector.AnimationUI.Color = function() { return "black"; }
    var timeline;
    InspectorTest.addSniffer(WebInspector.TabbedPane.prototype, "changeTabView", onChangeTabView, true);
    WebInspector.inspectorView.showViewInDrawer("animations", true);

    function onChangeTabView(id, view)
    {
        if (!timeline && id === "animations") {
            timeline = view;
            InspectorTest.assertTrue(timeline instanceof WebInspector.AnimationTimeline);
            InspectorTest.waitForAnimationAdded(step2);
            InspectorTest.evaluateInPage("startAnimationWithFillForwards()");
        }
    }

    function step2(group)
    {
        timeline._selectAnimationGroup(group);
        timeline._redraw();
        InspectorTest.addResult(">>>> Animation with forwards fill");
        InspectorTest.dumpAnimationTimeline(timeline);
        timeline._reset();
        InspectorTest.waitForAnimationAdded(step3);
        InspectorTest.evaluateInPage("startAnimationWithFillBackwards()");
    }

    function step3(group)
    {
        timeline._selectAnimationGroup(group);
        timeline._redraw();
        InspectorTest.addResult(">>>> Animation with backwards fill");
        InspectorTest.dumpAnimationTimeline(timeline);
        InspectorTest.waitForAnimationAdded(step4);
        InspectorTest.evaluateInPage("startAnimationWithFillBoth()");
    }

    function step4(group)
    {
        timeline._selectAnimationGroup(group);
        timeline._redraw();
        InspectorTest.addResult(">>>> Animation with both fill");
        InspectorTest.dumpAnimationTimeline(timeline);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the display of animations on the animation timeline with fill modes.
</p>

<div id="node" style="background-color: red; height: 100px"></div>

</body>
</html>
