third_party/WebKit/LayoutTests/inspector/animation/animation-timeline-fill-expected.txt

Tests the display of animations on the animation timeline with fill modes.

>>>> Animation with forwards fill
WebAnimation

<g style="transform: translateX(166.67px);"><g style="transform: translateX(0px);"><line class="animation-line" x1="7" y1="26" y2="26" x2="340.33" style="stroke: black;"></line><path class="animation-keyframe" d="M 0 26 L 0.02 10.85 L 0.70 10.85 L 2.64 10.85 L 5.74 10.85 L 9.96 10.85 L 15.21 10.85 L 21.44 10.85 L 28.58 10.85 L 36.55 10.85 L 45.29 10.85 L 54.74 10.85 L 64.83 10.85 L 75.49 10.85 L 86.65 10.85 L 98.25 10.85 L 110.21 10.85 L 122.48 10.85 L 134.98 10.85 L 147.65 10.85 L 160.42 10.85 L 173.22 10.85 L 185.99 10.85 L 198.66 10.85 L 211.15 10.85 L 223.42 10.85 L 235.37 10.85 L 246.96 10.85 L 258.11 10.85 L 268.76 10.85 L 278.83 10.85 L 288.27 10.85 L 296.99 10.85 L 304.95 10.85 L 312.06 10.85 L 318.27 10.85 L 323.50 10.85 L 327.69 10.85 L 330.77 10.85 L 332.67 10.85 L 333.33 10.85 L 333.33 26 Z" style="transform: translateX(7px); fill: black;"></path><circle class="animation-endpoint" cx="7.00" cy="26" r="3.5" style="stroke: black; fill: black;"></circle><circle class="animation-endpoint" cx="340.33" cy="26" r="3.5" style="stroke: black; fill: black;"></circle></g><g class="animation-tail-iterations"></g></g><line class="animation-delay-line" x1="7" y1="26" y2="26" x2="173.67" style="stroke: black;"></line><line class="animation-delay-line animation-fill" x1="7" y1="26" y2="26" x2="507.00" style="stroke: black; transform: translateX(500px);"></line>
>>>> Animation with backwards fill
WebAnimation

<g style="transform: translateX(166.67px);"><g style="transform: translateX(0px);"><line class="animation-line" x1="7" y1="26" y2="26" x2="340.33" style="stroke: black;"></line><path class="animation-keyframe" d="M 0 26 L 0.02 10.85 L 0.70 10.85 L 2.64 10.85 L 5.74 10.85 L 9.96 10.85 L 15.21 10.85 L 21.44 10.85 L 28.57 10.85 L 36.55 10.85 L 45.29 10.85 L 54.74 10.85 L 64.83 10.85 L 75.49 10.85 L 86.65 10.85 L 98.25 10.85 L 110.21 10.85 L 122.48 10.85 L 134.98 10.85 L 147.65 10.85 L 160.42 10.85 L 173.22 10.85 L 185.99 10.85 L 198.66 10.85 L 211.15 10.85 L 223.42 10.85 L 235.37 10.85 L 246.96 10.85 L 258.11 10.85 L 268.76 10.85 L 278.83 10.85 L 288.27 10.85 L 296.99 10.85 L 304.95 10.85 L 312.06 10.85 L 318.27 10.85 L 323.50 10.85 L 327.69 10.85 L 330.77 10.85 L 332.67 10.85 L 333.33 10.85 L 333.33 26 Z" style="transform: translateX(7px); fill: black;"></path><circle class="animation-endpoint" cx="7.00" cy="26" r="3.5" style="stroke: black; fill: black;"></circle><circle class="animation-endpoint" cx="340.33" cy="26" r="3.5" style="stroke: black; fill: black;"></circle></g><g class="animation-tail-iterations"></g></g><line class="animation-delay-line animation-fill" x1="7" y1="26" y2="26" x2="173.67" style="stroke: black;"></line><line class="animation-delay-line" x1="7" y1="26" y2="26" x2="173.67" style="stroke: black; transform: translateX(500px);"></line>
>>>> Animation with both fill
WebAnimation

<g style="transform: translateX(166.67px);"><g style="transform: translateX(0px);"><line class="animation-line" x1="7" y1="26" y2="26" x2="340.33" style="stroke: black;"></line><path class="animation-keyframe" d="M 0 26 L 0.02 10.85 L 0.70 10.85 L 2.64 10.85 L 5.74 10.85 L 9.96 10.85 L 15.21 10.85 L 21.44 10.85 L 28.58 10.85 L 36.55 10.85 L 45.29 10.85 L 54.74 10.85 L 64.83 10.85 L 75.49 10.85 L 86.65 10.85 L 98.25 10.85 L 110.21 10.85 L 122.48 10.85 L 134.98 10.85 L 147.65 10.85 L 160.42 10.85 L 173.22 10.85 L 185.99 10.85 L 198.66 10.85 L 211.15 10.85 L 223.42 10.85 L 235.37 10.85 L 246.96 10.85 L 258.11 10.85 L 268.76 10.85 L 278.83 10.85 L 288.27 10.85 L 296.99 10.85 L 304.95 10.85 L 312.06 10.85 L 318.27 10.85 L 323.50 10.85 L 327.69 10.85 L 330.77 10.85 L 332.67 10.85 L 333.33 10.85 L 333.33 26 Z" style="transform: translateX(7px); fill: black;"></path><circle class="animation-endpoint" cx="7.00" cy="26" r="3.5" style="stroke: black; fill: black;"></circle><circle class="animation-endpoint" cx="340.33" cy="26" r="3.5" style="stroke: black; fill: black;"></circle></g><g class="animation-tail-iterations"></g></g><line class="animation-delay-line animation-fill" x1="7" y1="26" y2="26" x2="173.67" style="stroke: black;"></line><line class="animation-delay-line animation-fill" x1="7" y1="26" y2="26" x2="507.00" style="stroke: black; transform: translateX(500px);"></line>
